requirements.txt
Flask==3.0.3
SQLAlchemy==2.0.35
PyMySQL==1.1.1
cryptography==43.0.1
python-dotenv==1.0.1
Flask-JWT-Extended==4.6.0
Flask-Cors==5.0.0
Flask-Caching==2.3.0
firebase-admin==6.5.0
google-cloud-secret-manager==2.20.2
google-cloud-sql-connector==1.12.1
gunicorn==20.1.0
alembic==1.13.3
mysqlclient==2.2.4

# Keep the rest of the dependencies as they are in the original requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.9
aiosignal==1.3.1
attrs==24.2.0
blinker==1.8.2
CacheControl==0.14.0
cachelib==0.9.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
frozenlist==1.4.1
gcloud==0.17.0
google-api-core==2.20.0
google-api-python-client==2.147.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.19.0
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
greenlet==3.1.1
grpc-google-iam-v1==0.13.1
grpcio==1.66.2
grpcio-status==1.66.2
httplib2==0.22.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
jws==0.1.3
Mako==1.3.5
MarkupSafe==2.1.5
msgpack==1.1.0
multidict==6.1.0
mysql-connector==2.2.9
oauth2client==3.0.0
proto-plus==1.24.0
protobuf==5.29.0rc1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pycryptodome==3.10.1
PyJWT==2.9.0
pyOpenSSL==24.2.1
pyparsing==3.1.4
python-jwt==2.0.1
redis==5.0.8
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9
setuptools==75.1.0
six==1.16.0
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
Werkzeug==3.0.4
wheel==0.44.0
yarl==1.13.1
